Refactor: CreatePotPage, DetailPage, MainPage 폴더 구조 변경
