+++ src/pages/DetailPage/components/modal/DeleteModal.tsx
