+++ src/pages/PotPage/index.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react';
 import instance from '@apis';
-import SingleContent from '../MainPage/Components/SingleContent';
 import { useMyPotContentStore } from '@stores/MyPotContentStore';
 import * as S from '../MainPage/style';
 import styled from 'styled-components';
 import { Chevronleft } from '@assets/svg';
 import { Icon } from '../DetailPage/style';
+import SingleContent from '@pages/MainPage/SingleContent';
 function PotPage() {
     const { id } = JSON.parse(localStorage.getItem('myInfo') as string);
     const { MyAppliedPotContent, MyPotContent, setMyPotContent, setMyAppliedPotContent, setTotalMyPotContent } =
